+++ buildSrc/src/main/java/Config.kt
@@ -1,12 +1,13 @@
 import java.math.BigDecimal
 
 object Config {
-    val kotlinVersion = "1.4.32"
+    val kotlinVersion = "1.5.31"
     val kotlinStdLib = "stdlib-jdk8"
 
     val springBootVersion = "2.4.4"
     // Spring is currently not compatible with Kotlin 1.4
     val springKotlinCompatibleLanguageVersion = "1.3"
+    val kotlinCompatibleLanguageVersion = "1.4"
 
     object BuildPlugins {
         val androidGradle = "com.android.tools.build:gradle:7.0.3"
@@ -97,6 +98,10 @@ object Config {
         val apolloCoroutines = "com.apollographql.apollo:apollo-coroutines-support:$apolloVersion"
 
         val p6spy = "p6spy:p6spy:3.9.1"
+
+        val graphQlJava = "com.graphql-java:graphql-java:17.3"
+
+        val kotlinReflect = "org.jetbrains.kotlin:kotlin-reflect"
     }
 
     object AnnotationProcessors {