+++ sentry/src/main/java/io/sentry/DateUtils.java
@@ -139,4 +139,35 @@ public final class DateUtils {
   public static @NotNull String getTimestampMinutesPrecision(final @NotNull Date date) {
     return SDF_ISO_FORMAT_MINUTES_PRECISION.get().format(date);
   }
+
+  /**
+   * Converts milliseconds to seconds.
+   *
+   * @param millis - milliseconds
+   * @return seconds
+   */
+  public static double millisToSeconds(final double millis) {
+    return millis / 1000;
+  }
+
+  /**
+   * Converts nanoseconds to milliseconds
+   *
+   * @param nanos - nanoseconds
+   * @return milliseconds
+   */
+  public static double nanosToMillis(final double nanos) {
+    return nanos / 1000000;
+  }
+
+  /**
+   * Convert {@link Date} to epoch time in seconds represented as {@link Double}.
+   *
+   * @param date - date
+   * @return seconds
+   */
+  @SuppressWarnings("JavaUtilDate")
+  public static double dateToSeconds(final @NotNull Date date) {
+    return millisToSeconds(date.getTime());
+  }
 }