+++ sentry/src/main/java/io/sentry/Hub.java
@@ -140,7 +140,7 @@ public final class Hub implements IHub {
         final StackItem item = stack.peek();
         sentryId = item.getClient().captureEvent(event, item.getScope(), hint);
         this.lastEventId = sentryId;
-      } catch (Exception e) {
+      } catch (Throwable e) {
         options
             .getLogger()
             .log(
@@ -166,7 +166,7 @@ public final class Hub implements IHub {
       try {
         final StackItem item = stack.peek();
         sentryId = item.getClient().captureMessage(message, level, item.getScope());
-      } catch (Exception e) {
+      } catch (Throwable e) {
         options.getLogger().log(SentryLevel.ERROR, "Error while capturing message: " + message, e);
       }
     }
@@ -194,7 +194,7 @@ public final class Hub implements IHub {
         if (capturedEnvelopeId != null) {
           sentryId = capturedEnvelopeId;
         }
-      } catch (Exception e) {
+      } catch (Throwable e) {
         options.getLogger().log(SentryLevel.ERROR, "Error while capturing envelope.", e);
       }
     }
@@ -219,7 +219,7 @@ public final class Hub implements IHub {
         final SentryEvent event = new SentryEvent(throwable);
         assignTraceContext(event);
         sentryId = item.getClient().captureEvent(event, item.getScope(), hint);
-      } catch (Exception e) {
+      } catch (Throwable e) {
         options
             .getLogger()
             .log(
@@ -259,7 +259,7 @@ public final class Hub implements IHub {
       try {
         final StackItem item = stack.peek();
         item.getClient().captureUserFeedback(userFeedback);
-      } catch (Exception e) {
+      } catch (Throwable e) {
         options
             .getLogger()
             .log(
@@ -316,7 +316,7 @@ public final class Hub implements IHub {
         if (sessionFlusher != null) {
           sessionFlusher.close();
         }
-      } catch (Exception e) {
+      } catch (Throwable e) {
         options.getLogger().log(SentryLevel.ERROR, "Error while closing the Hub.", e);
       }
       isEnabled = false;
@@ -506,7 +506,7 @@ public final class Hub implements IHub {
       pushScope();
       try {
         callback.run(stack.peek().getScope());
-      } catch (Exception e) {
+      } catch (Throwable e) {
         options.getLogger().log(SentryLevel.ERROR, "Error in the 'withScope' callback.", e);
       }
       popScope();
@@ -524,7 +524,7 @@ public final class Hub implements IHub {
     } else {
       try {
         callback.run(stack.peek().getScope());
-      } catch (Exception e) {
+      } catch (Throwable e) {
         options.getLogger().log(SentryLevel.ERROR, "Error in the 'configureScope' callback.", e);
       }
     }
@@ -557,7 +557,7 @@ public final class Hub implements IHub {
     } else {
       try {
         stack.peek().getClient().flush(timeoutMillis);
-      } catch (Exception e) {
+      } catch (Throwable e) {
         options.getLogger().log(SentryLevel.ERROR, "Error in the 'client.flush'.", e);
       }
     }
@@ -609,7 +609,7 @@ public final class Hub implements IHub {
             item = stack.peek();
             sentryId =
                 item.getClient().captureTransaction(transaction, traceState, item.getScope(), hint);
-          } catch (Exception e) {
+          } catch (Throwable e) {
             options
                 .getLogger()
                 .log(