+++ sentry/src/main/java/io/sentry/SentryClient.java
@@ -206,7 +206,7 @@ public final class SentryClient implements ISentryClient {
     for (final EventProcessor processor : eventProcessors) {
       try {
         event = processor.process(event, hint);
-      } catch (Exception e) {
+      } catch (Throwable e) {
         options
             .getLogger()
             .log(
@@ -237,7 +237,7 @@ public final class SentryClient implements ISentryClient {
     for (final EventProcessor processor : eventProcessors) {
       try {
         transaction = processor.process(transaction, hint);
-      } catch (Exception e) {
+      } catch (Throwable e) {
         options
             .getLogger()
             .log(
@@ -509,7 +509,7 @@ public final class SentryClient implements ISentryClient {
     if (beforeSend != null) {
       try {
         event = beforeSend.execute(event, hint);
-      } catch (Exception e) {
+      } catch (Throwable e) {
         options
             .getLogger()
             .log(