+++ sentry/src/test/java/io/sentry/SentryClientTest.kt
@@ -32,7 +32,6 @@ import java.io.IOException
 import java.io.InputStreamReader
 import java.lang.IllegalArgumentException
 import java.lang.IllegalStateException
-import java.lang.RuntimeException
 import java.nio.charset.Charset
 import java.util.Arrays
 import java.util.UUID
@@ -1195,7 +1194,7 @@ class SentryClientTest {
     private fun eventProcessorThrows(): EventProcessor {
         return object : EventProcessor {
             override fun process(event: SentryEvent, hint: Any?): SentryEvent? {
-                throw RuntimeException()
+                throw Throwable()
             }
         }
     }